add seal-hub-kit
