+++ inputs/input-balance.json
@@ -1,1 +1,1 @@
-{"address":"0xbf74483DB914192bb0a9577f3d8Fb29a6d4c08eE","addressPubKeyX":"4877328357478890623967823018480272757589824716691017530689013849938564609461","addressPubKeyY":"19319156333180214350448676801453385628019589905553133160599925263402925020311","addressR8x":"6131150531428173639876258616573452329317960670849539052997748214806084117872","addressR8y":"3542683127533152694423208573004577640359972623991635310709370226012721617971","addressS":"1552032294900630710671636126341686524618344690893907071544082165193326278704","balanceMessage":["0x00","0x25467bc5101e722a993cd81390c550a7239974dc73479fb399603a2e3f75cf69","0x722B0676F457aFe13e479eB2a8A4De88BA15B2c6","0x67","0x01"],"balancePubKeyX":"4877328357478890623967823018480272757589824716691017530689013849938564609461","balancePubKeyY":"19319156333180214350448676801453385628019589905553133160599925263402925020311","balanceR8x":"9553077237007568487335888732821279423066601742500511718204567877926834230448","balanceR8y":"17706048220997761681096159203563870746864837382163668528082297052933971273135","balanceS":"1147880556223594952111117911786143375779911404043293465718419604572472828907","pathIndices":[0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],"siblings":["0xc21cb669c1829c07aecbb985b223ec5f1172f88d","0x0f973e2684c634ce63dedf199a775ec5cfdde1a36ee05c72059c98308e2a9247","0x19f17d00def10c5dbbfbaf706d4a4e7af74ac34d71e6ea56bb4b3cbe06efed11","0x13a0ce73dee34e70f5ec1141e8355abb636235685e7c6671eaa0624ad0f110f2","0x1b7ac8e63b2a6148250232aa0404efafb2175af1dc69423aba17374d8f87244a","0x25bc300285aec2d33be2e119cee77a26005c77fa7c68400d21b4946eae547504","0x0733c6cb12b8fdfa3f8174970d2695ea148b0e96b8b150a828a24f1d15b424ea","0x2cd2236985d02cde4f4f82e8dc3f2d5a5b1e82f34fcd748a37516cb0c0359ab5","0x22202b644ce1794628d38e69d690c8cfbc0b9037f62448168ea8f77fcf502ff2","0x07a536a90267b7fa95f56b3c46ae14d30f7592adee4b7fd7c12f9e3967b5e315","0x23a6bfb0a95cd62f01ad4e0d208e380837b7f142c2e25c9945207db0ce6f74bd","0x1f307b83dadc0511955db1d61b6ac4fc75f5e640cef858f2a142de70bdb1dcae","0x1690cc667b699416b20d053f81441a740a87e8260661c6f10e90e9818afdcc43","0x02af12e3851d2423be8f56bfa21145d2c47c7fccfcf10cac2c58198f8599f53c","0x28c2a00784edec4b5dacb29db357d355c556d1f0f31f92dcfa21b943779515d2","0x1c6cc64e99a30764c384cd31b564c6ec157ae24ac975eb88980ae9720a9d91af","0x18c7ac6762fd1a97ed5655effe17bd8d5a4d58dc8a07abfe85ffb2e06a08748c","0x2408db1057565a46eb96e198da856eade8c96638d71d957eb097fc8b18083f0a","0x05f2887365f264a4bc9413831e81f8d14f8ea098dd922175ef97bc4440354296","0xfd96074536bfd500b27b0fdbe448072ec22c5ed7300f8c22186eed0af2e46a"],"nonce":["0x93cc32e6658175ac0d5f7d8256b09cdc768156216c462f3c4f6bf6094578b188","0xc8aa55fd688abd344a51bafe1ed8808357910d3f37874704ae7bba47be3691d4"]}
+{"r":[["12370936595211678540","13453180533785522820","10703475296946091640","6628273245074491717"]],"s":[["8382700564301861820","1345353828700683444","13548196630803365697","340490747005897074"]],"U":[["13219399089741314990","2478718603600695215","14212391934076437009","6874014124320662532"],["8158088631120511405","5539392206733686802","6786932440732189640","9333477952063626801"]],"pubKey":[["4543898144865705096","14300113079446097790","8584010577784086056","12703555293657500329"],["8810258805127543072","11110600409948878046","14966571377819271586","7482263276216251493"]],"pathIndices":[1,1,0,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],"siblings":["0x5a9a86b76337940f103babb8056d5cfcca5b8ba0c525d2db5070314745f8e28f","0xb7eab66fd20e8422c35346cab05135d49af8198f1da66023720ee17e8d2fe00b","0x50b4feaeb79752e57b182c6207a6984ebf5e6dc9d7e56c42889666509843b718","0xf56fdd59a3fd78fbc066b31c20a0dc02d2fab63095664e87f2b2f0819e1cc22d","0x6e58ea3b67b9d42ee340b22fcc79b87a8ce47a7a6d0404cb1d63fc16c0b95220","0x6a60e17ccdf3662131a55af3aab36767f12a57dee30fcac0c57a67c4ed73b62f","0x9d0fd75916dc9ba4e9a7bcd1946c5a0f351e0c8f4c8322dcfbc12435b1ac4601","0x4dd60b46e179bc509022284c4ba37c9992b2e1b4f3261480dc18c2b346a9a01c","0x4dc7695fdeb763e585c1fa1d235c42d196917acd8867cdcf20b5fca7594a3412","0x363f05d4d2cca7b40d87546181acd14f1d21f9535c3d13c45dfbb32afaa3c516","0xbeab72b4311584a18d104dbf69ef69690840fd9fc40263b58122052478f08117","0xe4f44df15cd40969d4f1bea1110ea66ba4e275ec3839ae243d72cd22f01f0d21","0xb159372c0d35324c8f5fe23ff3fdf89901218d3d544eafaa115c08f2ddf6e205","0xed736191e841bed7a395136f9fa614613debec5500f7ad6ef4d347ebdfd2dc03","0xcc5180e4ec4b20348de932af63145a84262d19dff70a7021004f908a3bbb0a14","0x794af051ca62f7b8442b34b6a502719aa31a49ba3abeda8186008fbb48f1331d","0x173f6cd904333bd65eff5ab013cdf034628a1a60c21e7cb9713b478716bddf15","0x1869e4f7fb4386284462475025833eb5dfeed358e6599fb341e5ddcaa077021f","0xec6e0d9aabf53541d3d255ea5822dcd43ead632aa2aebb0ba8f0ca3394b88e22","0x33fbba7f378f62141641f322f3f05d40161059883aeedbbd89f09388e3c70812","0x4459addee6aafc4ee7227ea42bbb89f483feee8523a9afa0916e5e8366897328","0xd7d0a92515d6f57edd30c2e9cfb11d1255a7f282d447074e720aad653c546d09","0x755aabf024d8210f6f3ebb5331688020231d58dcd430115b26fdd72e757ea107","0xead1caf46fb1368a7a5191f81728ed661cc4e371887ea6bf09c7da705cac7909","0x8e2a8d5f8ce53cbf47e2a5ae2a1a0e0dfcaecb3d25d7db303c444c72360cfd01","0x9984f3c2e03ac49515e87a9e02628a6077aee45c1e87de26b849e21a97a2912c","0x886a532698fc63ab729e65b481f2f63bf9774d8cb6b2e4d48de9faa88506751d","0x0b0ad4c0e5f7cf64acc7116ffb2c67f578da0eb69fe3d894bada1f800010b726","0x8c6a37733478e96d2c9cea21941b2a015310c6b1e30a5c1f3cb0da064b168f29","0xf330d4c7f25785032c21022c8bc7a47f523ef2181fe1274892098e0a2603872a"],"balanceMessage":["0x00","0x25467bc5101e722a993cd81390c550a7239974dc73479fb399603a2e3f75cf69","0x722B0676F457aFe13e479eB2a8A4De88BA15B2c6","0x67","0x01"],"balancePubKeyX":"4877328357478890623967823018480272757589824716691017530689013849938564609461","balancePubKeyY":"19319156333180214350448676801453385628019589905553133160599925263402925020311","balanceR8x":"9553077237007568487335888732821279423066601742500511718204567877926834230448","balanceR8y":"17706048220997761681096159203563870746864837382163668528082297052933971273135","balanceS":"1147880556223594952111117911786143375779911404043293465718419604572472828907","ownersPathIndices":[0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],"ownersSiblings":["0xc21cb669c1829c07aecbb985b223ec5f1172f88d","0x0f973e2684c634ce63dedf199a775ec5cfdde1a36ee05c72059c98308e2a9247","0x19f17d00def10c5dbbfbaf706d4a4e7af74ac34d71e6ea56bb4b3cbe06efed11","0x13a0ce73dee34e70f5ec1141e8355abb636235685e7c6671eaa0624ad0f110f2","0x1b7ac8e63b2a6148250232aa0404efafb2175af1dc69423aba17374d8f87244a","0x25bc300285aec2d33be2e119cee77a26005c77fa7c68400d21b4946eae547504","0x0733c6cb12b8fdfa3f8174970d2695ea148b0e96b8b150a828a24f1d15b424ea","0x2cd2236985d02cde4f4f82e8dc3f2d5a5b1e82f34fcd748a37516cb0c0359ab5","0x22202b644ce1794628d38e69d690c8cfbc0b9037f62448168ea8f77fcf502ff2","0x07a536a90267b7fa95f56b3c46ae14d30f7592adee4b7fd7c12f9e3967b5e315","0x23a6bfb0a95cd62f01ad4e0d208e380837b7f142c2e25c9945207db0ce6f74bd","0x1f307b83dadc0511955db1d61b6ac4fc75f5e640cef858f2a142de70bdb1dcae","0x1690cc667b699416b20d053f81441a740a87e8260661c6f10e90e9818afdcc43","0x02af12e3851d2423be8f56bfa21145d2c47c7fccfcf10cac2c58198f8599f53c","0x28c2a00784edec4b5dacb29db357d355c556d1f0f31f92dcfa21b943779515d2","0x1c6cc64e99a30764c384cd31b564c6ec157ae24ac975eb88980ae9720a9d91af","0x18c7ac6762fd1a97ed5655effe17bd8d5a4d58dc8a07abfe85ffb2e06a08748c","0x2408db1057565a46eb96e198da856eade8c96638d71d957eb097fc8b18083f0a","0x05f2887365f264a4bc9413831e81f8d14f8ea098dd922175ef97bc4440354296","0xfd96074536bfd500b27b0fdbe448072ec22c5ed7300f8c22186eed0af2e46a"]}
