+++ package.json
@@ -35,6 +35,7 @@
   },
   "devDependencies": {
     "@big-whale-labs/constants": "^0.1.35",
+    "@big-whale-labs/seal-hub-kit": "^0.0.3",
     "@big-whale-labs/seal-hub-verifier-template": "^0.1.0",
     "@big-whale-labs/versioned-contract": "^1.0.2",
     "@ethereumjs/util": "^8.0.2",
