--- utils/Input.ts
@@ -1,8 +1,0 @@
-export type BigIntOrString = bigint | string
-
-export interface Input {
-  U: BigIntOrString[][]
-  r: BigIntOrString[][]
-  s: BigIntOrString[][]
-  pubKey: BigIntOrString[][]
-}
