+++ utils/publicKeyToArraysSplitted.ts
@@ -1,0 +1,32 @@
+import BN from 'bn.js'
+
+const REGISTERS = 4n
+const addHexPrefix = (str: string) => `0x${str}`
+
+const splitToRegisters = (value?: BN | string) => {
+  const registers = [] as bigint[]
+
+  if (!value) {
+    return [0n, 0n, 0n, 0n]
+  }
+  const hex = value.toString('hex').padStart(64, '0')
+  for (let k = 0; k < REGISTERS; k++) {
+    // 64bit = 16 chars in hex
+    const val = hex.slice(k * 16, (k + 1) * 16)
+
+    registers.unshift(BigInt(addHexPrefix(val)))
+  }
+
+  return registers.map((el) => el.toString())
+}
+
+export default function publicKeyToArraysSplitted(publicKey: string) {
+  const x = splitToRegisters(
+    new BN(BigInt(addHexPrefix(publicKey.slice(4, 4 + 64))).toString())
+  )
+  const y = splitToRegisters(
+    new BN(BigInt(addHexPrefix(publicKey.slice(68, 68 + 64))).toString())
+  )
+
+  return [x, y]
+}
